restructured core module
