--- peepingtom/core/datacrate.py
@@ -1,4 +1,0 @@
-class DataCrate(list):
-    """
-    A container for DataBlock objects which exist within the same n-dimensional reference space
-    """
